Rename integration_blueprint to optispark
