+++ .github/ISSUE_TEMPLATE/bug.yml
@@ -22,7 +22,7 @@ body:
         required: true
       - label: This issue only contains 1 issue (if you have multiple issues, open one issue for each issue).
         required: true
-      - label: This issue is not a duplicate issue of any [previous issues](https://github.com/ludeeus/integration_blueprint/issues?q=is%3Aissue+label%3A%22Bug%22+)..
+      - label: This issue is not a duplicate issue of any [previous issues](https://github.com/Big-Tree/HomeAssistant-OptiSpark/issues?q=is%3Aissue+label%3A%22Bug%22+)..
         required: true
 - type: textarea
   attributes:
