+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -14,7 +14,7 @@ body:
         required: true
       - label: This only contains 1 feature request (if you have multiple feature requests, open one feature request for each feature request).
         required: true
-      - label: This issue is not a duplicate feature request of [previous feature requests](https://github.com/ludeeus/integration_blueprint/issues?q=is%3Aissue+label%3A%22Feature+Request%22+).
+      - label: This issue is not a duplicate feature request of [previous feature requests](https://github.com/Big-Tree/HomeAssistant-OptiSpark/issues?q=is%3Aissue+label%3A%22Feature+Request%22+).
         required: true
 
 - type: textarea
