+++ config/configuration.yaml
@@ -5,4 +5,4 @@ default_config:
 logger:
   default: info
   logs:
-    custom_components.integration_blueprint: debug
+    custom_components.optispark: debug
