+++ custom_components/optispark/__init__.py
@@ -1,7 +1,7 @@
-"""Custom integration to integrate integration_blueprint with Home Assistant.
+"""Custom integration to integrate optispark with Home Assistant.
 
 For more details about this integration, please refer to
-https://github.com/ludeeus/integration_blueprint
+https://github.com/Big-Tree/HomeAssistant-OptiSpark
 """
 from __future__ import annotations
