+++ custom_components/optispark/binary_sensor.py
@@ -1,0 +1,61 @@
+"""Binary sensor platform for optispark."""
+from __future__ import annotations
+
+from homeassistant.components.binary_sensor import (
+    BinarySensorDeviceClass,
+    BinarySensorEntity,
+    BinarySensorEntityDescription,
+)
+
+from .const import DOMAIN
+from .coordinator import BlueprintDataUpdateCoordinator
+from .entity import IntegrationBlueprintEntity
+
+ENTITY_DESCRIPTIONS = (
+    BinarySensorEntityDescription(
+        key="optispark",
+        name="Integration Blueprint Binary Sensor",
+        device_class=BinarySensorDeviceClass.CONNECTIVITY,
+    ),
+)
+
+
+async def async_setup_entry(hass, entry, async_add_devices):
+    """Set up the binary_sensor platform."""
+    coordinator = hass.data[DOMAIN][entry.entry_id]
+    async_add_devices(
+        IntegrationBlueprintBinarySensor(
+            coordinator=coordinator,
+            entity_description=entity_description,
+        )
+        for entity_description in ENTITY_DESCRIPTIONS
+    )
+
+
+class IntegrationBlueprintBinarySensor(IntegrationBlueprintEntity, BinarySensorEntity):
+    """optispark binary_sensor class."""
+
+    def __init__(
+        self,
+        coordinator: BlueprintDataUpdateCoordinator,
+        entity_description: BinarySensorEntityDescription,
+    ) -> None:
+        """Initialize the binary_sensor class."""
+        super().__init__(coordinator)
+        self.entity_description = entity_description
+
+    @property
+    def is_on(self) -> bool:
+        """Return true if the binary_sensor is on."""
+        return self.coordinator.data.get("title", "") == "foo"
+
+    @property
+    def unique_id(self):
+        return 'sensor_id_binary'
+
+    #@property
+    #def device_info(self):
+    #    """Return device information."""
+    #    return {
+    #        "identifiers": {("optispark_integration", "optispark_device_id")},
+    #    }
