+++ custom_components/optispark/const.py
@@ -1,9 +1,9 @@
-"""Constants for integration_blueprint."""
+"""Constants for optispark."""
 from logging import Logger, getLogger
 
 LOGGER: Logger = getLogger(__package__)
 
-NAME = "Integration blueprint"
-DOMAIN = "integration_blueprint"
+NAME = "Optispark"
+DOMAIN = "optispark"
 VERSION = "0.0.0"
 ATTRIBUTION = "Data provided by http://jsonplaceholder.typicode.com/"
