+++ custom_components/optispark/coordinator.py
@@ -1,4 +1,4 @@
-"""DataUpdateCoordinator for integration_blueprint."""
+"""DataUpdateCoordinator for optispark."""
 from __future__ import annotations
 
 from datetime import timedelta, datetime
