+++ custom_components/optispark/entity.py
