+++ custom_components/optispark/manifest.json
@@ -1,0 +1,13 @@
+{
+  "domain": "optispark",
+  "name": "Optispark",
+  "codeowners": [
+    "@Big-Tree",
+    "@chappers00"
+  ],
+  "config_flow": true,
+  "documentation": "https://github.com/Big-Tree/HomeAssistant-OptiSpark",
+  "iot_class": "cloud_polling",
+  "issue_tracker": "https://github.com/Big-Tree/HomeAssistant-OptiSpark/issues",
+  "version": "0.0.0"
+}
