+++ custom_components/optispark/number.py
@@ -1,4 +1,4 @@
-"""Sensor platform for integration_blueprint."""
+"""Sensor platform for optispark."""
 from __future__ import annotations
 
 from homeassistant.components.sensor import SensorEntity, SensorEntityDescription
@@ -11,7 +11,7 @@ from datetime import datetime
 
 ENTITY_DESCRIPTIONS = (
     SensorEntityDescription(
-        key="integration_blueprint",
+        key="optispark",
         name="Optimised Demand",
         icon="mdi:format-quote-close",
     ),
@@ -41,7 +41,7 @@ async def async_setup_entry(hass, entry, async_add_devices):
 
 
 class IntegrationBlueprintSensor(IntegrationBlueprintEntity, SensorEntity):
-    """integration_blueprint Sensor class."""
+    """optispark Sensor class."""
 
     def __init__(
         self,
