+++ custom_components/optispark/switch.py
@@ -1,30 +1,26 @@
-"""Binary sensor platform for integration_blueprint."""
+"""Switch platform for optispark."""
 from __future__ import annotations
 
-from homeassistant.components.binary_sensor import (
-    BinarySensorDeviceClass,
-    BinarySensorEntity,
-    BinarySensorEntityDescription,
-)
+from homeassistant.components.switch import SwitchEntity, SwitchEntityDescription
 
 from .const import DOMAIN
 from .coordinator import BlueprintDataUpdateCoordinator
 from .entity import IntegrationBlueprintEntity
 
 ENTITY_DESCRIPTIONS = (
-    BinarySensorEntityDescription(
-        key="integration_blueprint",
-        name="Integration Blueprint Binary Sensor",
-        device_class=BinarySensorDeviceClass.CONNECTIVITY,
+    SwitchEntityDescription(
+        key="optispark",
+        name="Integration Switch",
+        icon="mdi:format-quote-close",
     ),
 )
 
 
 async def async_setup_entry(hass, entry, async_add_devices):
-    """Set up the binary_sensor platform."""
+    """Set up the sensor platform."""
     coordinator = hass.data[DOMAIN][entry.entry_id]
     async_add_devices(
-        IntegrationBlueprintBinarySensor(
+        IntegrationBlueprintSwitch(
             coordinator=coordinator,
             entity_description=entity_description,
         )
@@ -32,26 +28,36 @@ async def async_setup_entry(hass, entry, async_add_devices):
     )
 
 
-class IntegrationBlueprintBinarySensor(IntegrationBlueprintEntity, BinarySensorEntity):
-    """integration_blueprint binary_sensor class."""
+class IntegrationBlueprintSwitch(IntegrationBlueprintEntity, SwitchEntity):
+    """optispark switch class."""
 
     def __init__(
         self,
         coordinator: BlueprintDataUpdateCoordinator,
-        entity_description: BinarySensorEntityDescription,
+        entity_description: SwitchEntityDescription,
     ) -> None:
-        """Initialize the binary_sensor class."""
+        """Initialize the switch class."""
         super().__init__(coordinator)
         self.entity_description = entity_description
 
     @property
     def is_on(self) -> bool:
-        """Return true if the binary_sensor is on."""
+        """Return true if the switch is on."""
         return self.coordinator.data.get("title", "") == "foo"
+
+    async def async_turn_on(self, **_: any) -> None:
+        """Turn on the switch."""
+        await self.coordinator.api.async_set_title("bar")
+        await self.coordinator.async_request_refresh()
+
+    async def async_turn_off(self, **_: any) -> None:
+        """Turn off the switch."""
+        await self.coordinator.api.async_set_title("foo")
+        await self.coordinator.async_request_refresh()
 
     @property
     def unique_id(self):
-        return 'sensor_id_binary'
+        return 'sensor_id_switch'
 
     #@property
     #def device_info(self):
