+++ requirements.txt
@@ -1,4 +1,4 @@
 colorlog==6.7.0
-homeassistant==2023.8.0
+homeassistant==2023.11.2
 pip>=21.0,<23.2
 ruff==0.0.292
